refactor: 플레이어 승패 결과에 따른 베팅 금액 계산
- WinningType은 수익 비율을 상태로 가지고 있으며,
multiplyProfitRate를 통해 입력 받은 금액에 수익 비율을
곱한 값을 반환한다.

- ProfitCalculator가 존재하는 이유는 이러한 수익 금액을
계산하는 로직이 WinningType도, WinningResultDiscriminator도
아니기 때문이다.
