+++ src/test/java/blackjack/view/WinningTypeTest.java
@@ -1,22 +1,18 @@
-package blackjack.domain;
+package blackjack.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.view.WinningType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class ProfitCalculatorTest {
-    @DisplayName("플레이어가_블랙잭_승리한_경우_수익금은_원금의_1.5배다")
+class WinningTypeTest {
+    @DisplayName("금액에_수익률을_곱해_반환할_수_있다")
     @CsvSource(value = {"BLACKJACK_WIN:15_000", "WIN:20_000", "DEFEAT:-10_000", "DRAW:10_000"}, delimiterString = ":")
     @ParameterizedTest
-    void calculatePlayerProfit(WinningType type, int expected) {
-        // given
-        ProfitCalculator profitCalculator = new ProfitCalculator();
-
+    void multiplyProfitRate(WinningType type, int expected) {
         // when
-        int result = profitCalculator.calculatePlayerProfit(type, 10_000);
+        int result = type.multiplyProfitRate(10_000);
 
         // then
         assertThat(result).isEqualTo(expected);
